+++ src/cli/default-error-handler.js
@@ -34,6 +34,7 @@ import {
   CorruptedComponent,
   VersionAlreadyExists,
   MergeConflict,
+  HashMismatch,
   MergeConflictOnRemote,
   VersionNotFound,
   CyclicDependencies
@@ -130,6 +131,13 @@ const errorsMap: Array<[Class<Error>, (err: Class<Error>) => string]> = [
   [RemoteNotFound, err => `error: remote "${chalk.bold(err.name)}" was not found`],
   [NetworkError, err => `error: remote failed with error the following error:\n "${chalk.bold(err.remoteErr)}"`],
   [
+    HashMismatch,
+    err => `found hash mismatch of ${chalk.bold(err.id)}, version ${chalk.bold(err.version)}.
+  originalHash: ${chalk.bold(err.originalHash)}.
+  currentHash: ${chalk.bold(err.currentHash)}
+  this usually happens when a component is old and the migration script was not running or interrupted`
+  ],
+  [
     MergeConflict,
     err =>
       `error: merge conflict occurred while importing the component ${err.id}. conflict version(s): ${err.versions.join(
@@ -154,7 +162,10 @@ to resolve this conflict and merge your remote and local changes, please do the 
 once your changes are merged with the new remote version, please tag and export a new version of the component to the remote scope.`
   ],
   [CyclicDependencies, err => `${err.msg.toString().toLocaleLowerCase()}`],
-  [UnexpectedNetworkError, () => 'error: unexpected network error has occurred'],
+  [
+    UnexpectedNetworkError,
+    err => `error: unexpected network error has occurred. ${err.message ? ` original message: ${err.message}` : ''}`
+  ],
   [SSHInvalidResponse, () => 'error: received an invalid response from the remote SSH server'],
   [ScopeNotFound, () => 'error: workspace not found. to create a new workspace, please use `bit init`'],
   [
@@ -313,6 +324,8 @@ export default (err: Error): ?string => {
 
   Analytics.setError(LEVEL.INFO, err.makeAnonymous());
   const [, func] = error;
-  logger.error(`User gets the following error: ${func(err)}`);
-  return chalk.red(func(err));
+  const errorMessage = func(err);
+  err.message = errorMessage;
+  logger.error(`User gets the following error: ${errorMessage}`);
+  return chalk.red(errorMessage);
 };