+++ src/consumer/login/login.js
@@ -6,11 +6,12 @@ import os from 'os';
 import chalk from 'chalk';
 import url from 'url';
 import { setSync, getSync } from '../../api/consumer/lib/global-config';
-import { CFG_USER_TOKEN_KEY, DEFAULT_HUB_LOGIN, CFG_HUB_LOGIN_KEY } from '../../constants';
+import { CFG_USER_TOKEN_KEY, DEFAULT_HUB_LOGIN, CFG_HUB_LOGIN_KEY, DEFAULT_LANGUAGE, DEFAULT_REGISTRY_URL, CFG_REGISTRY_URL_KEY } from '../../constants';
 import { LoginFailed } from '../exceptions';
 import logger from '../../logger/logger';
 import GeneralError from '../../error/general-error';
 import { Analytics } from '../../analytics/analytics';
+import { Driver } from '../../driver';
 
 const ERROR_RESPONSE = 500;
 const DEFAULT_PORT = 8085;
@@ -18,10 +19,13 @@ const REDIRECT = 302;
 
 export default function loginToBitSrc(
   port: string,
-  noLaunchBrowser?: boolean
+  noLaunchBrowser?: boolean,
+  npmrcPath: string,
+  skipRegistryConfig: boolean
 ): Promise<{ isAlreadyLoggedIn?: boolean, username?: string }> {
   return new Promise((resolve, reject) => {
     const clientGeneratedId = uuid();
+    const driver = Driver.load(DEFAULT_LANGUAGE);
     if (getSync(CFG_USER_TOKEN_KEY)) {
       // $FlowFixMe
       return resolve({ isAlreadyLoggedIn: true });
@@ -45,6 +49,7 @@ export default function loginToBitSrc(
           reject(new LoginFailed());
         }
         setSync(CFG_USER_TOKEN_KEY, token);
+        if (!skipRegistryConfig) driver.npmLogin(token, npmrcPath, getSync(CFG_REGISTRY_URL_KEY) || DEFAULT_REGISTRY_URL);
         response.writeHead(REDIRECT, { Location: redirectUri });
         closeConnection();
         resolve({ username });